+++ aeron-driver/src/main/c/aeron_publication_image.c
@@ -218,11 +218,11 @@ int aeron_publication_image_close(aeron_counters_manager_t *counters_manager, ae
 void aeron_publication_image_clean_buffer_to(aeron_publication_image_t *image, int64_t new_clean_position)
 {
     const int64_t clean_position = image->conductor_fields.clean_position;
-    const size_t bytes_for_cleaning = (size_t)(new_clean_position - clean_position);
+    const int32_t bytes_for_cleaning = (int32_t)(new_clean_position - clean_position);
     const size_t dirty_term_index = aeron_logbuffer_index_by_position(clean_position, image->position_bits_to_shift);
     const int32_t term_offset = (int32_t)(clean_position & image->term_length_mask);
-    const size_t bytes_left_in_term = image->term_length_mask + 1 - (size_t)term_offset;
-    const size_t length = bytes_for_cleaning < bytes_left_in_term ? bytes_for_cleaning : bytes_left_in_term;
+    const int32_t bytes_left_in_term = (int32_t)image->term_length_mask + 1 - term_offset;
+    const int32_t length = bytes_for_cleaning < bytes_left_in_term ? bytes_for_cleaning : bytes_left_in_term;
 
     if (length > 0)
     {
